PET-181 Add : Tab layout binding ver3
